pyproject.toml
[project]
name = "leather-store"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.1.2",
    "django>=6.0.3",
    "gunicorn>=25.3.0",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.12",
    "pytest-playwright>=0.7.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "ruff>=0.15.12",
    "whitenoise>=6.12.0",
]
[tool.uv]
link-mode = "copy"
